adding redirect to a profile page on login
