+++ src/App.js
@@ -4,10 +4,9 @@ import { Security, SecureRoute, ImplicitCallback } from '@okta/okta-react';
 import './App.css';
 import Navbar from './components/Navbar/navbar'
 import Home from './components/pages/Home'
-import MyTeams from './components/pages/MyTeams'
 import Login from './components/auth/login'
 import RegistrationForm from './components/auth/Registration/Registration'
-import ProfilePage from './components/auth/ProfilePage'
+import ProfilePage from './components/auth/ProfilePage';
 
 
 function onAuthRequired({history}) {
@@ -29,8 +28,7 @@ class App extends Component {
             <Navbar />
             <div className="container">
               <Route path='/' exact={true} component={Home} />
-              <SecureRoute path='/myteams' exact={true} component={MyTeams} />
-              <SecureRoute path="/profile" component={ProfilePage} />
+              <SecureRoute path='/profile' exact={true} component={ProfilePage} />
               <Route path='/login' render={() => <Login baseUrl='https://dev-451795.okta.com' />} />
               <Route path="/register" component={RegistrationForm} />
               <Route path='/implicit/callback' component={ImplicitCallback} />
