+++ src/components/Navbar/navbar.js
@@ -28,8 +28,8 @@ class Navbar extends Component {
                     </Link>
                   </li>
                   <li className="nav-item">
-                    <Link className="nav-link" to="/myteams">
-                      My Teams
+                    <Link className="nav-link" to="/profile">
+                      Profile
                     </Link>
                   </li>
                 </ul>
